Base Drag handling and position change working
