+++ Assets/Team/Prefabs/UI/UI-Game-Turn.prefab
@@ -13,7 +13,7 @@ GameObject:
   - component: {fileID: 7554951489642248323}
   - component: {fileID: 5501738679531262055}
   - component: {fileID: 6722041604154524934}
-  - component: {fileID: 461289132740429008}
+  - component: {fileID: 2588750479764618927}
   m_Layer: 5
   m_Name: UI-Game-Turn
   m_TagString: Untagged
@@ -110,7 +110,7 @@ MonoBehaviour:
   m_FlexibleWidth: -1
   m_FlexibleHeight: 0
   m_LayoutPriority: 1
---- !u!114 &461289132740429008
+--- !u!114 &2588750479764618927
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -119,6 +119,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 6258264882312573246}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9e27c9723f5c98c478d415e5f38193dc, type: 3}
+  m_Script: {fileID: 11500000, guid: 43496426aaa4c8d428d7b72a338f0db9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  originalParent: {fileID: 0}
