--- Assets/Team/Scripts/UI/DraggableUIElement.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9e27c9723f5c98c478d415e5f38193dc
